Compilation on Win64.
git-svn-id: http://www.bci2000.org/svn/trunk@4165 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/contrib/AppConnectorApplications/AppConnectorExample/src/mainUI.cpp
@@ -127,7 +127,7 @@ void mainUI::timerUpdate()
   QMutexLocker locker( &mMutex );
   if( states.size() != stateTable->rowCount() )
   {
-    stateTable->setRowCount( states.size() );
+    stateTable->setRowCount( static_cast<int>( states.size( )) );
     for( int row = 0; row < stateTable->rowCount(); ++row )
     {
       stateTable->setItem( row, 0, new QTableWidgetItem );
