+++ src/contrib/BCPy2000/foundation/shared/PythonFilter.cpp
@@ -221,15 +221,15 @@ FILTER_NAME::FILTER_NAME()
     PyObject *py_states = PyTuple_GetItem(py_ret, 1);
 
     // Add the parameters and states requested by the Python _Construct function.
-    int numParamDefs = PyList_Size(py_params);
-    for( int i = 0; i < numParamDefs; ++i ) {
+    size_t numParamDefs = PyList_Size(py_params);
+    for( size_t i = 0; i < numParamDefs; ++i ) {
       string paramDef = PyString_AsString(PyList_GetItem(py_params, i));
       // std::cerr << "registering parameter: " << paramDef << std::endl;
       if(!Parameters->Add( paramDef ) ) DoubleErr(("Error in parameter definition: " + paramDef).c_str(), "Construct");
     }
 
-    int numStateDefs = PyList_Size(py_states);
-    for( int i = 0; i < numStateDefs; ++i ) {
+    size_t numStateDefs = PyList_Size(py_states);
+    for( size_t i = 0; i < numStateDefs; ++i ) {
       string stateDef = PyString_AsString(PyList_GetItem(py_states, i));
       // std::cerr << "registering state: " << stateDef << std::endl;
       if(!States->Add( stateDef )) DoubleErr(("Error in state definition: " + stateDef).c_str(), "Construct");
@@ -295,8 +295,8 @@ FILTER_NAME::Preflight( const SignalProperties& inSignalProperties,
     PyObject* py_list;
     py_list = PyObject_GetAttrString(bci2000_instance, (char*)"_writeable_params");
     if(py_list && PyList_Check(py_list)) {
-      int n = PyList_Size(py_list);
-      for(int i = 0; i < n; i++) {
+      size_t n = PyList_Size(py_list);
+      for(size_t i = 0; i < n; i++) {
         const char *name = PyString_AsString(PyList_GetItem(py_list, i));
         Parameter(name); // allows write access to this parameter
       }
