+++ src/core/Operator/OperatorQt/MainWindow.h
@@ -92,7 +92,7 @@ class MainWindow : public QMainWindow
   void GetParameters();
 
  public slots:
-  void SetFunctionButton( int idx, const QString& title, const QString& script );
+  void SetFunctionButton( size_t idx, const QString& title, const QString& script );
   void PutParameters();
 
  private:
